:truck: Rename ContextInfra file
