+++ data_safe_haven/context/__init__.py
@@ -1,4 +1,4 @@
-from .context import ContextInfra
+from .context_infra import ContextInfra
 from .context_settings import Context, ContextSettings
 
 __all__ = [
